requirements.txt
anyio==3.6.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.95.1
Flask==2.3.2
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gspread==5.8.0
h11==0.14.0
halo==0.0.31
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
log-symbols==0.0.14
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
plotly==5.14.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
spinners==0.0.24
starlette==0.26.1
tenacity==8.2.2
termcolor==2.3.0
textblob==0.17.1
tqdm==4.65.0
tweepy==4.14.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
Werkzeug==2.3.3
